Bugfix: no principal in Gateway domain, but was trying to sign
